Moved fieldsValue state and some functions from MainView.jsx to AddProject.jsx to cleaner code.
